Phase 2: Break up large storage module into focused components
Storage module refactoring complete:
- storage/mod.rs: 1503 → 28 lines (98% reduction)
- Extracted storage/backend.rs: StorageBackend trait (85 lines)
- Extracted storage/types.rs: Data types and structures (97 lines)
- Extracted storage/memory.rs: InMemoryStorage implementation (1,138 lines)
- Extracted storage/manager.rs: StorageManager coordination (224 lines)

Benefits:
✅ Single-responsibility modules for better maintainability
✅ Clear API boundaries between storage components
✅ Easier testing with co-located tests
✅ Better code organization and navigation
✅ Fixed type compatibility issues
